+++ plugins/GetSocialFacebookInvitePlugin.m
@@ -39,7 +39,7 @@
     self.cancelCallback = cancelCallback;
 
     FBSDKAppInviteContent *content = [[FBSDKAppInviteContent alloc] init];
-    if (invitePackage.imageUrl)
+    if(invitePackage.imageUrl && invitePackage.image)
     {
         NSURL* appImageUrl = [NSURL URLWithString:invitePackage.imageUrl];
         if (appImageUrl != nil)