+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## Version History
 
+### v6.2.0
+
+**Added**:
++ Added `ActivityDetailsView`, now you can open activity comments from code.
++ Added `executeWhenInitialized` method to be able to execute any action right after SDK is initialized.
++ Added `setAvatarClickHandler` method to `GetSocialUIActivityFeedView` and `GetSocialUIActivityDetailsView`. You can use this method to get notified if a user clicked on a user's avatar.
++ Added `image` property to `GetSocialMutableInviteContent` to set invite image using binaries.
++ Facebook and Kakao plugins share image as well in Smart Invites.
+
+
+
+**Fixed**
++ Custom translated text issue in Smart Invites.
+
+
+
+----
+
 ### v6.1.0
 
 **Fixed**: